refactor: move lib/postal/smtp_server to app/lib/smtp_server
