--- lib/postal/smtp_server.rb
@@ -1,6 +1,0 @@
-# frozen_string_literal: true
-
-module Postal
-  module SMTPServer
-  end
-end
