+++ script/smtp_server.rb
@@ -1,4 +1,4 @@
 # frozen_string_literal: true
 
 require_relative "../config/environment"
-Postal::SMTPServer::Server.new(debug: true).run
+SMTPServer::Server.new(debug: true).run
