--- spec/lib/postal/smtp_server/client_spec.rb
@@ -1,14 +1,0 @@
-# frozen_string_literal: true
-
-require "rails_helper"
-
-module Postal
-  module SMTPServer
-
-    describe Client do
-      let(:ip_address) { "1.2.3.4" }
-      subject(:client) { described_class.new(ip_address) }
-    end
-
-  end
-end
